requirements.txt
# frappe -- https://github.com/frappe/frappe
# erpnext -- https://github.com/frappe/erpnext
cryptography>=41.0.0
requests>=2.31.0
python-dateutil>=2.8.2
xmltodict>=0.13.0
